Add 04-cyklus

+++ 01-uvod/03-vypocty.py
@@ -1,5 +1,5 @@
-a = int(input("Zadaj číslo a: ")) # Input je príkazom vstupu. Vypíše nám na obrazovku text v zátvorke a to, čo užívateľ napíše na klávesnici, uloží do premennej a
-b = int(input("Zadaj číslo b: ")) # Input vracia vždy výsledok vo formáte reťazca (postupnosť znakov resp. text), funkcia int() pretypuje reťazecna celé číslo
+a = float(input("Zadaj číslo a: ")) # Input je príkazom vstupu. Vypíše nám na obrazovku text v zátvorke a to, čo užívateľ napíše na klávesnici, uloží do premennej a
+b = float(input("Zadaj číslo b: ")) # Input vracia vždy výsledok vo formáte reťazca (postupnosť znakov resp. text), funkcia int() pretypuje reťazecna celé číslo
 sucet = a + b
 print("Súčet je: ", sucet)
 rozdiel = a - b
